+++ cmd/pipelines/templates/updates.go
@@ -1,4 +1,4 @@
-package output
+package templates
 
 import (
 	"context"
